requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
async-timeout==4.0.2
attrs==23.1.0
banal==1.0.6
certifi==2023.5.7
charset-normalizer==3.1.0
colorlog==6.7.0
dataset==1.6.0
frozenlist==1.3.3
hikari==2.0.0.dev119
hikari-lightbulb==2.3.3
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
python-dotenv==1.0.0
requests==2.30.0
SQLAlchemy==1.4.48
typing_extensions==4.5.0
urllib3==2.0.2
uvloop==0.17.0
yarl==1.9.2
